% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat4MLM2TE_obj.r
\name{dat4MLM2TE_obj}
\alias{dat4MLM2TE_obj}
\title{Utility function dat4MLM2TE_obj}
\usage{
dat4MLM2TE_obj(dat, cutoff = 0)
}
\arguments{
\item{dat}{dataframe with names y, site, species, trait and env
for values of abundance, identity of site and of species and the trait value and environmental value, respectively.}

\item{cutoff}{minimal number of occurrences of species}
}
\value{
An object of class TE_obj
}
\description{
\code{dat4MLM2TE_obj} creates from data in a format for glm(m)/MLM a \code{TE_obj} object (see \code{\link{make_obj_for_traitenv}}), which consists of
trait, environment and abundance data. The argument \code{dat} is usually created using
\code{\link{expand4glmm}}. The function is used repeatedly in \code{\link{MLM3_p_max}}.
}
\details{
The order of data in dat should be in standard order, that is, first all data from site 1, then site 2,
with all species listed in identical order in each site (including zero abundances).
BEWARE: The function currently works for a single trait and environmental variable only!!!!
}
\examples{
\dontrun{
data("Revisit")
str(Revisit)
TE_obj <- dat4MLM2TE_obj(Revisit)
str(TE_obj)
}
}
\references{
ter Braak (2019) New robust weighted averaging- and model-based methods
for assessing trait-environment relationships. Methods in Ecology and Evolution (https://doi.org/10.1111/2041-210X.13278)
}
