% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MLM3.r
\name{plot_MLM3}
\alias{plot_MLM3}
\title{Plotting the interaction of a MLM3 model}
\usage{
plot_MLM3(MLM3, verbose = FALSE,
  title = paste("site and species effects (fixed + random) against env and trait"))
}
\arguments{
\item{MLM3}{the fitted MLM3 object, created by glmer (lme4) or glmmTMB.}

\item{verbose}{logical; default FALSE. If TRUE, a summary of the MLM3 object is printed.}

\item{title}{character text to adapt the main title.}
}
\value{
A list of printable and modifiable objects of class ggplot2.
}
\description{
\code{plot_MLM3} plots the trait-environment interaction
of a fitted MLM3 object.
}
\details{
The code uses libraries ggplot2 and dplyr.
}
\examples{
\dontrun{
data("Revisit")
formula.MLM3 <- y ~  env*trait  + (1 + env|species) + (1 + trait| site)
MLM3 <- glmmTMB(formula.MLM3, family = betabinomial,  data=Revisit)
plot(MLM3)
}
}
\references{
ter Braak (2019) New robust weighted averaging- and model-based methods
for assessing trait-environment relationships. Methods in Ecology and Evolution (https://doi.org/10.1111/2041-210X.13278)
}
\seealso{
\code{\link{expand4glmm}}.
}
